feat(blog/helper): centralise post parsing for wider use
